backend/requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
django-prometheus==2.4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
prometheus_client==0.24.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
uritemplate==4.2.0
whitenoise==6.11.0
